+++ android-app/app/src/main/java/com/echopen/asso/echopen/MainActivity.java
@@ -15,6 +15,7 @@ import android.view.animation.RotateAnimation;
 import android.widget.ImageView;
 import android.animation.ObjectAnimator;
 import android.view.animation.Animation;
+import android.os.Handler;
 
 
 import com.echopen.asso.echopen.echography_image_streaming.EchographyImageStreamingService;
@@ -35,7 +36,7 @@ import com.echopen.asso.echopen.utils.Constants;
  * These two methods should be refactored into one
  */
 
-public class MainActivity extends Activity implements EchographyImageVisualisationContract.View {
+public class MainActivity extends Activity {
 
     private static final String TAG = MainActivity.class.getSimpleName();
     private EchographyImageStreamingService mEchographyImageStreamingService;
@@ -46,7 +47,7 @@ public class MainActivity extends Activity implements EchographyImageVisualisati
     private ImageView mBatteryButton;
     private ImageView mSelectButton;
     private ImageView mCaptureShadow;
-
+    private Handler handler;
 
     private Long then;
     private RotateAnimation rotate_animation_capture;
@@ -74,90 +75,9 @@ public class MainActivity extends Activity implements EchographyImageVisualisati
 
         setContentView(R.layout.activity_main);
 
-        mCaptureButton = (ImageView) findViewById(R.id.main_button_capture);
-
-        mCaptureShadow = (ImageView) findViewById(R.id.main_button_shadow);
-        mCaptureShadow.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Log.d("captureButton", "Short Press");
 
 
-            }
-        });
-        mCaptureShadow.setOnTouchListener(new View.OnTouchListener() {
-            @Override
 
-            public boolean onTouch(View v,final MotionEvent event) {
-                if(event.getAction() == MotionEvent.ACTION_DOWN){
-                    then = System.currentTimeMillis();
-                   mEchographyImageVisualisationPresenter.toggleFreeze();
-
-                    rotate_animation_capture = new RotateAnimation(0,144 ,
-                            Animation.RELATIVE_TO_SELF, 0.5f, Animation.RELATIVE_TO_SELF,
-                            0.6f);
-
-                    rotate_animation_capture.setDuration(5000);
-                    mCaptureShadow.clearAnimation();
-                    mCaptureShadow.setAnimation(rotate_animation_capture);
-
-                }
-
-
-                else if(event.getAction() == MotionEvent.ACTION_UP){
-                    mEchographyImageVisualisationPresenter.toggleFreeze();
-                    if(((Long) System.currentTimeMillis() - then) > 5000){
-                        Log.d("mcaptureButton", "Long Press");
-
-                        return true;
-                    }
-                    else {
-
-                        rotate_animation_capture.cancel();
-                        rotate_animation_capture= null;
-
-                    }
-                }
-                return false;
-            }
-        });
-
-
-        mPregnantWomanButton = (ImageView) findViewById(R.id.main_button_mode);
-        mPregnantWomanButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Log.d("pregnantButton", "PregnantWomanButton Pressed");
-            }
-        });
-
-        mSelectButton = (ImageView) findViewById(R.id.main_button_select);
-        mSelectButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Log.d("selectButton", "SelectButton Pressed");
-            }
-        });
-
-        mEndExamButton = (ImageView) findViewById(R.id.main_button_end_exam);
-        mEndExamButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Log.d("endExamButton", "EndExamButton Pressed");
-            }
-        });
-
-        mBatteryButton = (ImageView) findViewById(R.id.main_button_battery);
-        mBatteryButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Log.d("batteryButton", "BatteryButton Pressed");
-            }
-        });
-
-
-        mEchographyImageStreamingService.getRenderingContextController().setLinearLutSlope(RenderingContext.DEFAULT_LUT_SLOPE);
-        mEchographyImageStreamingService.getRenderingContextController().setLinearLutOffset(RenderingContext.DEFAULT_LUT_OFFSET);
     }
 
     /**
@@ -178,38 +98,4 @@ public class MainActivity extends Activity implements EchographyImageVisualisati
 
 
 
-    @Override
-    public void refreshImage(final Bitmap iBitmap) {
-        this.runOnUiThread(new Runnable() {
-            @Override
-            public void run() {
-                Log.d(TAG, "refreshImage");
-                ImageView lEchOpenImage = (ImageView) findViewById(R.id.echopenImage);
-                lEchOpenImage.setRotation(IMAGE_ROTATION_FACTOR);
-                lEchOpenImage.setScaleX(IMAGE_ZOOM_FACTOR);
-                lEchOpenImage.setScaleY(IMAGE_ZOOM_FACTOR);
-                lEchOpenImage.setImageBitmap(iBitmap);
-            }
-        });
     }
-
-    @Override
-   public void displayFreezeButton() {
-        mCaptureShadow.setImageResource(R.drawable.icon_arc_shadow);
-        mCaptureButton.setImageResource(R.drawable.button_jauge);
-
-    }
-
-    @Override
-    public void displayUnfreezeButton() {
-        mCaptureShadow.setImageResource(R.drawable.icon_save_image);
-
-
-    }
-
-    @Override
-    public void setPresenter(EchographyImageVisualisationContract.Presenter iPresenter) {
-        mEchographyImageVisualisationPresenter = iPresenter;
-        mEchographyImageVisualisationPresenter.start();
-    }
-}
