+++ android-app/app/src/main/res/layout/fragment_validation.xml
@@ -1,0 +1,31 @@
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context="com.echopen.asso.echopen.ValidationFragment">
+
+
+    />
+
+    <ImageButton
+
+        android:id="@+id/bouton_validation"
+        android:layout_width="80dp"
+        android:layout_height="80dp"
+        android:layout_gravity="bottom|left"
+        android:layout_marginBottom="50dp"
+        android:scaleType="fitCenter"
+        android:src="@drawable/btn_yes" />
+
+    <ImageButton
+
+        android:id="@+id/bouton_rejet"
+        android:layout_width="80dp"
+        android:layout_height="80dp"
+        android:layout_gravity="bottom|right"
+        android:layout_marginBottom="50dp"
+        android:scaleType="fitCenter"
+        android:src="@drawable/btn_no" />
+
+
+</FrameLayout>
